refactor: move vote delegate sync logic into own hook
